% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beard.R
\name{get_beard}
\alias{get_beard}
\title{get_beard}
\usage{
get_beard(width, height)
}
\arguments{
\item{width}{numeric}

\item{height}{numeric}
}
\value{
link character
}
\description{
Gets the link containing a picture of a bearded man
}
\examples{
get_beard(480, 640)
}

