% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corgi.R
\name{get_corgi}
\alias{get_corgi}
\title{get_corgi}
\usage{
get_corgi(width, height)
}
\arguments{
\item{width}{numeric}

\item{height}{numeric}
}
\value{
link
}
\description{
Inserts the link of a picture containing corgis  depending on the height and width specified
}
\examples{
get_corgi(300, 400)
}

