% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kitty.R
\name{get_kitty}
\alias{get_kitty}
\title{get_kitty}
\usage{
get_kitty(width, height)
}
\arguments{
\item{width}{numeric}

\item{height}{numeric}
}
\value{
link character
}
\description{
Places the link of a cat picture when specifying the width and height
}
\examples{
get_kitty(400, 500)
}

