% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random.R
\name{get_random}
\alias{get_random}
\title{get_random}
\usage{
get_random(width, height)
}
\arguments{
\item{width}{numeric}

\item{height}{numeric}
}
\value{
link
}
\description{
Gets the link of a random animal
}
\examples{
get_random(300, 300)
}

