% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zombies.R
\name{get_zombies}
\alias{get_zombies}
\title{get_zombies}
\usage{
get_zombies(width, height)
}
\arguments{
\item{width}{numeric}

\item{height}{numeric}
}
\value{
character
}
\description{
Gets a link that contains the picture of zombies with the specified height and width
}
\examples{
get_zombies(500, 600)
}

