% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmeme.R
\name{getmeme}
\alias{getmeme}
\title{get meme}
\usage{
getmeme(x)
}
\arguments{
\item{x}{words for the meme generator}
}
\value{
doge meme
}
\description{
Gets the link for a doge meme in which you can change the words. Just input three words and the meme will be generated
}
\examples{
getmeme(c("wow", "nice", "function"))
}

