% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbacon.R
\name{getbacon}
\alias{getbacon}
\title{getbacon}
\usage{
getbacon(x)
}
\arguments{
\item{x}{numeric}
}
\value{
character
}
\description{
Places the link for a random picture of bacon when specifying the height and width
}

