% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbears.R
\name{getbears}
\alias{getbears}
\title{getbears}
\usage{
getbears(x)
}
\arguments{
\item{x}{numeric}
}
\value{
character
}
\description{
Places the link for a random picture of a bear or bears when specifying the height and width
}
\examples{
getbears(500)
}

